Rename errors util to sharedMessages
