+++ packages/apollo/src/commands/client/check.ts
@@ -8,7 +8,7 @@ import { relative } from "path";
 import { graphqlTypes } from "apollo-language-server";
 import chalk from "chalk";
 import envCi from "env-ci";
-import { graphUndefinedError } from "../../utils/errors";
+import { graphUndefinedError } from "../../utils/sharedMessages";
 
 const { ValidationErrorType } = graphqlTypes;
 type ValidationResult = graphqlTypes.ValidateOperations_service_validateOperations_validationResults;
